[CHORE] UIImage 익스텐션 옮기기 (#33)
