+++ AMaDda/Global/Extension/UIImage+Extension.swift
@@ -1,0 +1,33 @@
+//
+//  UIImage+Extension.swift
+//  AMaDda
+//
+//  Created by 이성호 on 2022/07/18.
+//
+
+import UIKit
+
+extension UIImage {
+    static func load(name: String) -> UIImage {
+        guard let image = UIImage(named: name, in: nil, compatibleWith: nil) else {
+            return UIImage()
+        }
+        image.accessibilityIdentifier = name
+        return image
+    }
+    
+    static func load(systemName: String) -> UIImage {
+        guard let image = UIImage(systemName: systemName, compatibleWith: nil) else {
+            return UIImage()
+        }
+        image.accessibilityIdentifier = systemName
+        return image
+    }
+    
+    func resize(to size: CGSize) -> UIImage {
+        let image = UIGraphicsImageRenderer(size: size).image { _ in
+            draw(in: CGRect(origin: .zero, size: size))
+        }
+        return image
+    }
+}
