+++ AMaDda/Global/Literal/ImageLiteral.swift
@@ -15,28 +15,3 @@ enum ImageLiterals {
     // MARK: - button
     static var btnProfile: UIImage { .load(systemName: "person.circle.fill") }
 }
-
-extension UIImage {
-    static func load(name: String) -> UIImage {
-        guard let image = UIImage(named: name, in: nil, compatibleWith: nil) else {
-            return UIImage()
-        }
-        image.accessibilityIdentifier = name
-        return image
-    }
-    
-    static func load(systemName: String) -> UIImage {
-        guard let image = UIImage(systemName: systemName, compatibleWith: nil) else {
-            return UIImage()
-        }
-        image.accessibilityIdentifier = systemName
-        return image
-    }
-    
-    func resize(to size: CGSize) -> UIImage {
-        let image = UIGraphicsImageRenderer(size: size).image { _ in
-            draw(in: CGRect(origin: .zero, size: size))
-        }
-        return image
-    }
-}
