+++ AMaDda/Global/Supports/SceneDelegate.swift
@@ -18,7 +18,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let windowScene = (scene as? UIWindowScene) else { return }
          window = UIWindow(windowScene: windowScene)
-         let mainVC = AddingViewController()
+         let mainVC = MainViewController()
          window?.rootViewController = mainVC
          window?.makeKeyAndVisible()
     }
